+++ src/nop_video_decoder.cpp
@@ -2,6 +2,7 @@
 
 // WebRTC
 #include <api/video/i420_buffer.h>
+#include <media/base/media_constants.h>
 #include <modules/video_coding/codecs/av1/libaom_av1_encoder.h>
 #include <modules/video_coding/codecs/h264/include/h264.h>
 #include <modules/video_coding/codecs/vp8/include/vp8.h>
@@ -27,7 +28,7 @@ int32_t NopVideoDecoder::Decode(const webrtc::EncodedImage& input_image,
           .set_video_frame_buffer(i420_buffer)
           .set_timestamp_rtp(input_image.RtpTimestamp())
           .build();
-  callback_->Decoded(decoded_image, absl::nullopt, absl::nullopt);
+  callback_->Decoded(decoded_image, std::nullopt, std::nullopt);
 
   return WEBRTC_VIDEO_CODEC_OK;
 }
@@ -65,11 +66,12 @@ NopVideoDecoderFactory::GetSupportedFormats() const {
   for (const webrtc::SdpVideoFormat& format : h264_codecs) {
     supported_codecs.push_back(format);
   }
+  supported_codecs.push_back(webrtc::SdpVideoFormat(cricket::kH265CodecName));
   return supported_codecs;
 }
 
-std::unique_ptr<webrtc::VideoDecoder>
-NopVideoDecoderFactory::CreateVideoDecoder(
+std::unique_ptr<webrtc::VideoDecoder> NopVideoDecoderFactory::Create(
+    const webrtc::Environment& env,
     const webrtc::SdpVideoFormat& format) {
   return std::unique_ptr<webrtc::VideoDecoder>(
       absl::make_unique<NopVideoDecoder>());